pyproject.toml
[tool.poetry]
name = "domaincatch"
version = "0.1.0"
description = ""
authors = ["Denis Maslennikov <ktotom7@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pysocks = "^1.7.1"
beautifulsoup4 = "^4.12.3"
lxml = "^5.3.0"
requests = {extras = ["socks"], version = "^2.32.3"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
